+++ CHANGELOG.md
@@ -4,6 +4,7 @@ This is the changelog of our "JTL WooCommerce Connector".
 
 ## Unreleased
 - CO-2875 - use the correct \InvalidArgumentException
+- CO-2909 - Fix array_map error in germanized fields
 - CO-2888 - add default values for title and alt text
 
 ## 2.3.2 _2025-02-24_